requirements.txt
# requirements.txt (Versión para Gemini)
Flask>=2.0
google-generativeai>=0.5.0